requirements.txt
aniso8601==4.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-RESTful==0.3.7
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
line-bot-sdk==1.8.0
MarkupSafe==1.1.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
pymongo==3.7.2
gunicorn==19.9.0
dnspython==1.16.0